+++ logs/CHANGELOG.md
@@ -2,6 +2,9 @@
 ## 2026-05-12 17:31 | auto: 05-12 17:31 | src/styles/theme.ts 
 - src/styles/theme.ts
 
+## 2026-05-12 17:31 | auto: 05-12 17:31 | src/styles/theme.ts 
+- src/styles/theme.ts
+
 ## 2026-05-12 17:30 | auto: 05-12 17:30 | src/styles/theme.ts 
 - src/styles/theme.ts